requirements.txt
Django
djangorestframework
drf-yasg
django-filter
